Update + prepare for driverless printing
This patch modifies the cups packages,
so it is ready to be a driverless printing proxy.
This is especially useful for iOS and android.
most packages have been linked to DBUS,
so a link with avahi (mDNS) could be made

1) Update a few packages to be up-to-date
2) get rid of deprecated static archives
3) add necessary quirks to enable driverless printing
ghostscript needs to be enabled for postscript IPP to work,
otherwise the gsto* filters need to be removed manually:
rm /usr/lib/cups/filters/gsto*

Tested using:
1. HPCups with some modifications compared to hpcups in openwrt repos
(I'm still thinking how to pack them)
2. HP Envy 4526 (which accepts pcl3gui2, and needs hpcups plugin
3. pdf printing via ipp filtered locally via pdftoraster -> hpcups
4. Android 15 default printing engine

Signed-off-by jpsollie <janpieter.sollie@edpnet.be>